move count-context to utils
